[REFACTOR] moved TableWidget out os SpecViewer
